requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
matplotlib>=3.7.0
